+++ src/ts/components/uicontainer.ts
@@ -159,7 +159,6 @@ export class UIContainer extends Container<UIContainerConfig> {
     }, {
       name: 'focusin',
       handler: (e) => {
-        console.warn(e);
         showUi();
       },
     }, {