+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Web Content Accessibility (WCAG) 2.0
 - Slider controls for Web Content Accessibility
+- Fixed issue where settings panel show subscription was ignored on small UI screens
 
 ### Fixed
 - UI flickering back and forth on live streams after ads